tools/automatic_battery_tester/requirements.txt
argcomplete==3.6.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.4
idna==3.10
ifaddr==0.2.0
inquirerpy==0.3.4
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.3.1
packaging==25.0
pandas==2.3.0
pfzy==0.3.4
pillow==11.2.1
prompt_toolkit==3.0.51
psutil==7.0.0
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.0
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyVISA==1.15.0
PyVISA-py==0.8.0
requests==2.32.4
scipy==1.16.0
six==1.17.0
toml==0.10.2
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
zeroconf==0.147.0
